chore: move cookie constants to a separate file
