+++ src/constants/cookie.ts
@@ -1,0 +1,2 @@
+export const AUTHORIZATION = 'Authorization'
+export const RECENT_MAP_ID = 'recent_map_id'
